+++ lib/tasks/install/preact.rb
@@ -1,0 +1,3 @@
+@ruby2js_options = {filters: ['react']}
+@yarn_add='preact'
+eval IO.read "#{__dir__}/webpacker.rb"